app4/requirements.txt
# App4: Video Commercial Generator - Additional Requirements

# MongoDB
motor==3.3.2
pymongo==4.6.1

# Existing requirements from main project
google-genai
fastapi
uvicorn[standard]
python-dotenv
requests
botocore
plotly
python-multipart

# Additional for App4
pydantic==2.5.3
python-multipart
aiofiles

# WebAI-to-API Integration
aiohttp>=3.9.0
